update deploy script

+++ deploy.sh
@@ -1,26 +1,47 @@
 #!/bin/bash
 
+if ! command -v poetry &> /dev/null; then
+    echo "You need to install poetry to run this script"
+    exit
+fi
 
-pip install pytest
+if ! command -v poetry run pytest &> /dev/null; then
+    echo "Installing pytest"
+    poetry add pytest
+fi
 
 echo "testing ..."
 
-pytest ./cloud_logger
-pytest ./cloudlog_commons
-pytest ./lambda/read_logs
-pytest ./lambda/save_logs
-pytest ./logger_daemon
+poetry run pytest ./cloud_logger
+poetry run pytest ./cloudlog_commons
+poetry run pytest ./lambda/read_logs
+poetry run pytest ./lambda/save_logs
+poetry run pytest ./logger_daemon
 
 echo "Testing Successfull !"
 
 set -e
 
+if ! command -v terraform &> /dev/null; then
+    echo "The 'terraform' is not installed. Install it to continue the deployment"
+    exit
+fi
+
+if ! command -v aws &> /dev/null; then
+    echo "The 'aws'/'aws-cli' is not installed. Install it to continue the deployment"
+    exit
+fi
+
 cd infra
+terraform validate
 # terraform apply
 
 echo "Infrastrucutre deployment done"
 
-cd ../log_sniffer/build
+cd ../log_sniffer
+npm run build
+
+cd ./build
 aws s3 rm --recursive s3://sniffer.cloudlog.com/*
 aws s3 cp --recursive ./. s3://sniffer.cloudlog.com
